+++ forui/lib/forui.dart
@@ -39,4 +39,5 @@ export 'widgets/switch.dart';
 export 'widgets/tabs.dart';
 export 'widgets/text_field.dart';
 export 'widgets/tile.dart';
+export 'widgets/time_picker.dart';
 export 'widgets/tooltip.dart';