+++ forui_hooks/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.10.0 (Next)
 
+* Add `useFTimePickerController`.
 * **Breaking** Change `useFDatePickerController` to `useFDateFieldController`.
-
 
 ## 0.9.0